Move weights from double to float (#4)
* updates
* added EncodedEmbeddingLayer
* improvements
* fixes

+++ Ametrin.Numerics/GlobalUsings.cs
@@ -1,3 +1,3 @@
-global using Weight = double;
-global using SimdVector = System.Numerics.Vector<double>;
+global using Weight = float;
+global using SimdVector = System.Numerics.Vector<float>;
 global using SimdVectorHelper = System.Numerics.Vector;
